JWT Auth

+++ routes/tweet.py
@@ -15,10 +15,10 @@ from fastapi import Response
 from fastapi import status
 from fastapi import Body
 from fastapi import Path
+from fastapi import Depends
 
 # Models
 from models import Tweet
-from models import User
 
 # Database
 from config.db import connection
@@ -26,17 +26,20 @@ from config.db import connection
 # Schemas
 from schemas.tweet import Tweet as TweetOut
 from schemas.tweet import TweetWithRelations
-from schemas.tweet import RegisterTweet
+from schemas.tweet import BaseTweet
+from schemas.user import User as UserSchema
 
+# Middlewares
+from middleware.auth import get_current_user
 
 router = APIRouter()
 
 
-@router.get('/tweets/',
-         response_model=List[TweetWithRelations],
-         status_code=status.HTTP_200_OK,
-         summary='Get all tweets',
-         tags=['Tweets'])
+@router.get('/',
+            response_model=List[TweetWithRelations],
+            status_code=status.HTTP_200_OK,
+            summary='Get all tweets',
+            tags=['Tweets'])
 def list_tweets():
     """List tweets.
 
@@ -81,11 +84,11 @@ def list_tweets():
     return output
 
 
-@router.get('/tweets/{id}',
-         response_model=TweetWithRelations,
-         status_code=status.HTTP_200_OK,
-         summary='Get a tweet',
-         tags=['Tweets'])
+@router.get('/{id}',
+            response_model=TweetWithRelations,
+            status_code=status.HTTP_200_OK,
+            summary='Get a tweet',
+            tags=['Tweets'])
 def retrieve_tweet(
     id: int = Path(...,
                    title='Tweet ID',
@@ -138,12 +141,15 @@ def retrieve_tweet(
     return pydottie.transform(tweet)
 
 
-@router.post('/tweets/',
+@router.post('/',
           response_model=TweetOut,
           status_code=status.HTTP_200_OK,
           summary='Create a new tweet',
           tags=['Tweets'])
-def create_tweet(tweet: RegisterTweet = Body(...)):
+def create_tweet(
+    tweet: BaseTweet = Body(...),
+    request_user: UserSchema = Depends(get_current_user),
+):
     """Creates a tweet.
 
     This path operation creates a new tweet in the app.
@@ -160,15 +166,9 @@ def create_tweet(tweet: RegisterTweet = Body(...)):
     - user_id: **int**
     """
 
-    # Check if the user exists
-    user = connection.execute(User.select().where(User.c.id == tweet.user_id)).fetchone()
-
-    if user is None:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
-                            detail='User not found')
-
     # Create tweet
     tweet_dict = tweet.dict()
+    tweet_dict['user_id'] = request_user.id
 
     response = connection.execute(Tweet.insert().values(**tweet_dict))
 
@@ -183,7 +183,7 @@ def create_tweet(tweet: RegisterTweet = Body(...)):
     return tweet_dict
 
 
-@router.put('/tweets/{id}',
+@router.put('/{id}',
          response_model=TweetWithRelations,
          status_code=status.HTTP_200_OK,
          summary='Update tweet',
@@ -192,15 +192,18 @@ def update_tweet(
     id: str = Path(...,
                    title='Tweet ID',
                    description='The ID of the tweet to update'),
-    tweet: RegisterTweet = Body(...),
+    tweet: BaseTweet = Body(...),
+    request_user: UserSchema = Depends(get_current_user),
 ):
     """Update tweet.
 
     This path operation allows to update the content of a tweet.
 
+    Only the tweet owner can update the tweet.
+
     Parameters:
     - Request body parameters:
-        - tweet: **RegisterTweet**
+        - tweet: **BaseTweet**
 
     Returns a json with the updated tweet information:
     - id: **int**
@@ -210,42 +213,26 @@ def update_tweet(
     - user: **UserOut**
     """
 
-    response = connection.execute(Tweet.update().where(Tweet.c.id == id).values(**tweet.dict()))
+    tweet_response = connection.execute(Tweet.select().where(Tweet.c.id == id)).fetchone()
 
-    if response.rowcount == 0:
+    if tweet_response is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail='Tweet not found')
 
-    query = """
-    SELECT
-        t.id as 'id',
-        t.content as 'content',
-        t.created_at as 'created_at',
-        t.updated_at as 'updated_at',
-        u.id as 'user.id',
-        u.first_name as 'user.first_name',
-        u.last_name as 'user.last_name',
-        u.birth_date as 'user.birth_date',
-        u.email as 'user.email',
-        u.created_at as 'user.created_at',
-        u.updated_at as 'user.updated_at'
-    FROM
-        tweets as t
-    INNER JOIN
-        users as u
-    ON
-        t.user_id = u.id
-    WHERE
-        t.id = :id
-    ;
-    """
+    if tweet_response.user_id != request_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail='You are not allowed to update this tweet')
 
-    tweet = connection.execute(text(query), id=id).fetchone()
+    connection.execute(Tweet.update().where(Tweet.c.id == id).values(**tweet.dict()))
 
-    return pydottie.transform(tweet)
+    tweet_dict = {**tweet_response}
+    tweet_dict['updated_at'] = datetime.utcnow()
+    tweet_dict['user'] = request_user
+
+    return tweet_dict
 
 
-@router.delete('/tweets/{id}',
+@router.delete('/{id}',
             status_code=status.HTTP_204_NO_CONTENT,
             summary='Delete tweet',
             tags=['Tweets'])
@@ -254,20 +241,29 @@ def delete_tweet(
                    gt=0,
                    title='Tweet ID',
                    description='The ID of the tweet to delete'),
+    request_user: UserSchema = Depends(get_current_user),
 ):
     """Delete tweet.
 
     This path operation deletes a tweet in the app.
+
+    Only the tweet owner can delete the tweet.
 
     Parameters:
     - Request body parameters:
         - **tweet: RegisterTweet**
     """
 
-    response = connection.execute(Tweet.delete().where(Tweet.c.id == id))
+    tweet_response = connection.execute(Tweet.select().where(Tweet.c.id == id)).fetchone()
 
-    if response.rowcount == 0:
+    if tweet_response is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail='Tweet not found')
+
+    if tweet_response.user_id != request_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail='You are not allowed to delete this tweet')
+
+    connection.execute(Tweet.delete().where(Tweet.c.id == id))
 
     return Response(status_code=status.HTTP_204_NO_CONTENT)
